+++ src/app/layout.tsx
@@ -1,36 +1,27 @@
 import "@/styles/globals.css";
 
-import { type Metadata } from "next";
-import { Geist } from "next/font/google";
-
-import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
-import { AppSidebar } from "@/components/app-sidebar";
+import {type Metadata} from "next";
+import {Geist} from "next/font/google";
 
 export const metadata: Metadata = {
-  title: "BizLenz",
-  description: "Business proposals, at your finest.",
-  icons: [{ rel: "icon", url: "/favicon.ico" }],
+    title: "BizLenz",
+    description: "Business proposals, at your finest.",
+    icons: [{rel: "icon", url: "/favicon.ico"}],
 };
 
 const geist = Geist({
-  subsets: ["latin"],
-  variable: "--font-geist-sans",
+    subsets: ["latin"],
+    variable: "--font-geist-sans",
 });
 
 export default function RootLayout({
-  children,
-}: Readonly<{ children: React.ReactNode }>) {
-  return (
-    <html lang="en" className={`dark ${geist.variable}`}>
-      <body>
-        <SidebarProvider>
-          <AppSidebar />
-          <main>
-            <SidebarTrigger />
-            {children}
-          </main>
-        </SidebarProvider>
-      </body>
-    </html>
-  );
+                                       children,
+                                   }: Readonly<{ children: React.ReactNode }>) {
+    return (
+        <html lang="en" className={`dark ${geist.variable}`}>
+        <body>
+        {children}
+        </body>
+        </html>
+    );
 }
